# Make calculating overlapping areas faster