--- a/app/broadcast_areas/repo.py
+++ b/app/broadcast_areas/repo.py
@@ -141,6 +141,23 @@ class BroadcastAreasRepository(object):
 
         return areas
 
+    def get_areas_with_simple_polygons(self, area_ids):
+        q = """
+        SELECT broadcast_areas.id, name, count_of_phones, broadcast_area_library_id, simple_polygons
+        FROM broadcast_areas
+        JOIN broadcast_area_polygons on broadcast_area_polygons.id = broadcast_areas.id
+        WHERE broadcast_areas.id IN ({})
+        """.format(("?," * len(area_ids))[:-1])
+
+        results = self.query(q, *area_ids)
+
+        areas = [
+            (row[0], row[1], row[2], row[3], json.loads(row[4]))
+            for row in results
+        ]
+
+        return areas
+
     def get_all_areas_for_library(self, library_id):
         is_multi_tier_library = self.query("""
         SELECT exists(