--- a/app/models/broadcast_message.py
+++ b/app/models/broadcast_message.py
@@ -18,9 +18,9 @@ from app.notify_client.broadcast_message_api_client import (
     broadcast_message_api_client,
 )
 
-ESTIMATED_AREA_OF_LARGEST_UK_COUNTY = broadcast_area_libraries.get_areas(
+ESTIMATED_AREA_OF_LARGEST_UK_COUNTY = broadcast_area_libraries.get_areas([
     'ctyua19-E10000023'  # North Yorkshire
-)[0].polygons.estimated_area
+])[0].polygons.estimated_area
 
 
 class BroadcastMessage(JSONModel):
@@ -203,7 +203,7 @@ class BroadcastMessage(JSONModel):
 
     def get_areas(self, areas):
         return broadcast_area_libraries.get_areas(
-            *areas
+            areas
         )
 
     def get_simple_polygons(self, areas):