At the moment, when looking at how much an arbitrary polygon overlaps known areas, we intersect the arbitrary polygon with the full resolution polygon of each potential known area.

By using the simplified polygons instead of the full resolutions ones we:
- query less data from SQLite
- pass less data around
- give Shapely a less complicated shape to do its intersection and calculations on

This makes it faster to calculate how much of each electoral ward a custom area overlaps.

We trade off a small amount of accuracy by doing this, but nothing significant considering that we round the numbers before displaying them to the user.

For larger polygons, which overlap many electoral wards we can also improve performance by using a `JOIN` to reduce the number of SQLite queries we are doing from _n_ to 1 (where _n_ is the number of potential overlapping areas).

It also turns out we were doing all of the above multiple times because `BroadcastMessage.count_of_phones_likely` appeared multiple times in the Jinja template. So we get some extra benefit by using [`werkzeug.cached_property`](https://tedboy.github.io/flask/generated/generated/werkzeug.cached_property.html) to make sure this is only calculated once per page view.

# Results

For just the part of the code which calculates the count of likely phones:

Place represented by custom area | Time to calculate number of phones | Time to calculate number of phones using `simple_polygons` | Time to calculate number of phones using `simple_polygons` and with query optimisation
---------------------------------|--------|--------| ----
Bristol                          | 0.07s  | 0.02s | 0.01s
Skye                             | 0.02s  | 0.01s | 0.01s 


To render a whole page with a large polygon from our library (which uses the overlap of the bleed area to count the number of likely phones):

Library polygon | Time to render preview page before | Time to render preview page after | Time to render preview page with caching
---|---|---|---
Lincolnshire | 1.01s | 0.76s | 0.32s